:book: Explain how cmake-re tooling can be integrated as ccache remote_storage alternative
This explains how ccache can store cache hits on a Bazel RE-API cluster,
through the use of CCACHE_PREFIX as well as how it can be complemented to
support additionally caching of archiving and linking steps.

This also adds a page explaining how to integrate a custom build system
that is not normally supported, the mileage and details of the cluster
settings might vary because of execution platforms and labels.

However the ccache approach has been extensively tested and should always
work as it only requires actionCacheLookup, downloadBlob, uploadBlob types
of actions.